Code for Hang Test
I couldn't use the pulleyTestBranch because there is the exception code on it and I didn't think it would work with this code because when I pulled it it had a lot of errors. I made OldTestBranch incorrectly and was too afraid to delete it so its there with unusable code.